move matplotlib import inside ImageToTensorBoard (#1399)
* move matplotlib import inside ImageToTensorBoard class so that it is not a hard dependency for GPflow (`import gpflow` does not require matplotlib, only instantiating `ImageToTensorBoard` does)
* split up monitor into base.py and tensorboard.py
* add matplotlib to extras_require
